chore/cleanup: split API file up
the file is incredible huge, and really hard to grasp. In order to follow up on
\#373 I plan the following changes:

1. split up api (this commit).
2. introduce interfaces for specific areas which interact with the cache {servers, images, bootscripts, volumes, security groups, snapshots}
3. introduce api struct which implements api interface, but handles caching
4. modify CLI to use caching decorator around API, remove caching from API implementation
