+++ pkg/api/api.go
@@ -1102,73 +1102,6 @@ func (s *ScalewayAPI) handleHTTPError(goodStatusCode []int, resp *http.Response)
 	return body, nil
 }
 
-func (s *ScalewayAPI) fetchServers(api string, query url.Values, out chan<- ScalewayServers) func() error {
-	return func() error {
-		resp, err := s.GetResponsePaginate(api, "servers", query)
-		if err != nil {
-			return err
-		}
-		defer resp.Body.Close()
-
-		body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-		if err != nil {
-			return err
-		}
-		var servers ScalewayServers
-
-		if err = json.Unmarshal(body, &servers); err != nil {
-			return err
-		}
-		out <- servers
-		return nil
-	}
-}
-
-// GetServers gets the list of servers from the ScalewayAPI
-func (s *ScalewayAPI) GetServers(all bool, limit int) (*[]ScalewayServer, error) {
-	query := url.Values{}
-	if !all {
-		query.Set("state", "running")
-	}
-	if limit > 0 {
-		// FIXME: wait for the API to be ready
-		// query.Set("per_page", strconv.Itoa(limit))
-		panic("Not implemented yet")
-	}
-	if all && limit == 0 {
-		s.Cache.ClearServers()
-	}
-	var (
-		g    errgroup.Group
-		apis = []string{
-			ComputeAPIPar1,
-			ComputeAPIAms1,
-		}
-	)
-
-	serverChan := make(chan ScalewayServers, 2)
-	for _, api := range apis {
-		g.Go(s.fetchServers(api, query, serverChan))
-	}
-
-	if err := g.Wait(); err != nil {
-		return nil, err
-	}
-	close(serverChan)
-	var servers ScalewayServers
-
-	for server := range serverChan {
-		servers.Servers = append(servers.Servers, server.Servers...)
-	}
-
-	for i, server := range servers.Servers {
-		servers.Servers[i].DNSPublic = server.Identifier + URLPublicDNS
-		servers.Servers[i].DNSPrivate = server.Identifier + URLPrivateDNS
-		s.Cache.InsertServer(server.Identifier, server.Location.ZoneID, server.Arch, server.Organization, server.Name)
-	}
-	return &servers.Servers, nil
-}
-
 // ScalewaySortServers represents a wrapper to sort by CreationDate the servers
 type ScalewaySortServers []ScalewayServer
 
@@ -1184,694 +1117,6 @@ func (s ScalewaySortServers) Less(i, j int) bool {
 	date1, _ := time.Parse("2006-01-02T15:04:05.000000+00:00", s[i].CreationDate)
 	date2, _ := time.Parse("2006-01-02T15:04:05.000000+00:00", s[j].CreationDate)
 	return date2.Before(date1)
-}
-
-// GetServer gets a server from the ScalewayAPI
-func (s *ScalewayAPI) GetServer(serverID string) (*ScalewayServer, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, "servers/"+serverID, url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-
-	var oneServer ScalewayOneServer
-
-	if err = json.Unmarshal(body, &oneServer); err != nil {
-		return nil, err
-	}
-	// FIXME arch, owner, title
-	oneServer.Server.DNSPublic = oneServer.Server.Identifier + URLPublicDNS
-	oneServer.Server.DNSPrivate = oneServer.Server.Identifier + URLPrivateDNS
-	s.Cache.InsertServer(oneServer.Server.Identifier, oneServer.Server.Location.ZoneID, oneServer.Server.Arch, oneServer.Server.Organization, oneServer.Server.Name)
-	return &oneServer.Server, nil
-}
-
-// PostServerAction posts an action on a server
-func (s *ScalewayAPI) PostServerAction(serverID, action string) error {
-	data := ScalewayServerAction{
-		Action: action,
-	}
-	resp, err := s.PostResponse(s.computeAPI, fmt.Sprintf("servers/%s/action", serverID), data)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = s.handleHTTPError([]int{http.StatusAccepted}, resp)
-	return err
-}
-
-// DeleteServer deletes a server
-func (s *ScalewayAPI) DeleteServer(serverID string) error {
-	defer s.Cache.RemoveServer(serverID)
-	resp, err := s.DeleteResponse(s.computeAPI, fmt.Sprintf("servers/%s", serverID))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if _, err = s.handleHTTPError([]int{http.StatusNoContent}, resp); err != nil {
-		return err
-	}
-	return nil
-}
-
-// PostServer creates a new server
-func (s *ScalewayAPI) PostServer(definition ScalewayServerDefinition) (string, error) {
-	definition.Organization = s.Organization
-
-	resp, err := s.PostResponse(s.computeAPI, "servers", definition)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusCreated}, resp)
-	if err != nil {
-		return "", err
-	}
-	var server ScalewayOneServer
-
-	if err = json.Unmarshal(body, &server); err != nil {
-		return "", err
-	}
-	// FIXME arch, owner, title
-	s.Cache.InsertServer(server.Server.Identifier, server.Server.Location.ZoneID, server.Server.Arch, server.Server.Organization, server.Server.Name)
-	return server.Server.Identifier, nil
-}
-
-// PatchUserSSHKey updates a user
-func (s *ScalewayAPI) PatchUserSSHKey(UserID string, definition ScalewayUserPatchSSHKeyDefinition) error {
-	resp, err := s.PatchResponse(AccountAPI, fmt.Sprintf("users/%s", UserID), definition)
-	if err != nil {
-		return err
-	}
-
-	defer resp.Body.Close()
-
-	if _, err := s.handleHTTPError([]int{http.StatusOK}, resp); err != nil {
-		return err
-	}
-	return nil
-}
-
-// PatchServer updates a server
-func (s *ScalewayAPI) PatchServer(serverID string, definition ScalewayServerPatchDefinition) error {
-	resp, err := s.PatchResponse(s.computeAPI, fmt.Sprintf("servers/%s", serverID), definition)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if _, err := s.handleHTTPError([]int{http.StatusOK}, resp); err != nil {
-		return err
-	}
-	return nil
-}
-
-// PostSnapshot creates a new snapshot
-func (s *ScalewayAPI) PostSnapshot(volumeID string, name string) (string, error) {
-	definition := ScalewaySnapshotDefinition{
-		VolumeIDentifier: volumeID,
-		Name:             name,
-		Organization:     s.Organization,
-	}
-	resp, err := s.PostResponse(s.computeAPI, "snapshots", definition)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusCreated}, resp)
-	if err != nil {
-		return "", err
-	}
-	var snapshot ScalewayOneSnapshot
-
-	if err = json.Unmarshal(body, &snapshot); err != nil {
-		return "", err
-	}
-	// FIXME arch, owner, title
-	s.Cache.InsertSnapshot(snapshot.Snapshot.Identifier, "", "", snapshot.Snapshot.Organization, snapshot.Snapshot.Name)
-	return snapshot.Snapshot.Identifier, nil
-}
-
-// PostImage creates a new image
-func (s *ScalewayAPI) PostImage(volumeID string, name string, bootscript string, arch string) (string, error) {
-	definition := ScalewayImageDefinition{
-		SnapshotIDentifier: volumeID,
-		Name:               name,
-		Organization:       s.Organization,
-		Arch:               arch,
-	}
-	if bootscript != "" {
-		definition.DefaultBootscript = &bootscript
-	}
-
-	resp, err := s.PostResponse(s.computeAPI, "images", definition)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusCreated}, resp)
-	if err != nil {
-		return "", err
-	}
-	var image ScalewayOneImage
-
-	if err = json.Unmarshal(body, &image); err != nil {
-		return "", err
-	}
-	// FIXME region, arch, owner, title
-	s.Cache.InsertImage(image.Image.Identifier, "", image.Image.Arch, image.Image.Organization, image.Image.Name, "")
-	return image.Image.Identifier, nil
-}
-
-// PostVolume creates a new volume
-func (s *ScalewayAPI) PostVolume(definition ScalewayVolumeDefinition) (string, error) {
-	definition.Organization = s.Organization
-	if definition.Type == "" {
-		definition.Type = "l_ssd"
-	}
-
-	resp, err := s.PostResponse(s.computeAPI, "volumes", definition)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusCreated}, resp)
-	if err != nil {
-		return "", err
-	}
-	var volume ScalewayOneVolume
-
-	if err = json.Unmarshal(body, &volume); err != nil {
-		return "", err
-	}
-	// FIXME: s.Cache.InsertVolume(volume.Volume.Identifier, volume.Volume.Name)
-	return volume.Volume.Identifier, nil
-}
-
-// PutVolume updates a volume
-func (s *ScalewayAPI) PutVolume(volumeID string, definition ScalewayVolumePutDefinition) error {
-	resp, err := s.PutResponse(s.computeAPI, fmt.Sprintf("volumes/%s", volumeID), definition)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = s.handleHTTPError([]int{http.StatusOK}, resp)
-	return err
-}
-
-// ResolveServer attempts to find a matching Identifier for the input string
-func (s *ScalewayAPI) ResolveServer(needle string) (ScalewayResolverResults, error) {
-	servers, err := s.Cache.LookUpServers(needle, true)
-	if err != nil {
-		return servers, err
-	}
-	if len(servers) == 0 {
-		if _, err = s.GetServers(true, 0); err != nil {
-			return nil, err
-		}
-		servers, err = s.Cache.LookUpServers(needle, true)
-	}
-	return servers, err
-}
-
-// ResolveVolume attempts to find a matching Identifier for the input string
-func (s *ScalewayAPI) ResolveVolume(needle string) (ScalewayResolverResults, error) {
-	volumes, err := s.Cache.LookUpVolumes(needle, true)
-	if err != nil {
-		return volumes, err
-	}
-	if len(volumes) == 0 {
-		if _, err = s.GetVolumes(); err != nil {
-			return nil, err
-		}
-		volumes, err = s.Cache.LookUpVolumes(needle, true)
-	}
-	return volumes, err
-}
-
-// ResolveSnapshot attempts to find a matching Identifier for the input string
-func (s *ScalewayAPI) ResolveSnapshot(needle string) (ScalewayResolverResults, error) {
-	snapshots, err := s.Cache.LookUpSnapshots(needle, true)
-	if err != nil {
-		return snapshots, err
-	}
-	if len(snapshots) == 0 {
-		if _, err = s.GetSnapshots(); err != nil {
-			return nil, err
-		}
-		snapshots, err = s.Cache.LookUpSnapshots(needle, true)
-	}
-	return snapshots, err
-}
-
-// ResolveImage attempts to find a matching Identifier for the input string
-func (s *ScalewayAPI) ResolveImage(needle string) (ScalewayResolverResults, error) {
-	images, err := s.Cache.LookUpImages(needle, true)
-	if err != nil {
-		return images, err
-	}
-	if len(images) == 0 {
-		if _, err = s.GetImages(); err != nil {
-			return nil, err
-		}
-		images, err = s.Cache.LookUpImages(needle, true)
-	}
-	return images, err
-}
-
-// ResolveBootscript attempts to find a matching Identifier for the input string
-func (s *ScalewayAPI) ResolveBootscript(needle string) (ScalewayResolverResults, error) {
-	bootscripts, err := s.Cache.LookUpBootscripts(needle, true)
-	if err != nil {
-		return bootscripts, err
-	}
-	if len(bootscripts) == 0 {
-		if _, err = s.GetBootscripts(); err != nil {
-			return nil, err
-		}
-		bootscripts, err = s.Cache.LookUpBootscripts(needle, true)
-	}
-	return bootscripts, err
-}
-
-// GetImages gets the list of images from the ScalewayAPI
-func (s *ScalewayAPI) GetImages() (*[]MarketImage, error) {
-	images, err := s.GetMarketPlaceImages("")
-	if err != nil {
-		return nil, err
-	}
-	s.Cache.ClearImages()
-	for i, image := range images.Images {
-		if image.CurrentPublicVersion != "" {
-			for _, version := range image.Versions {
-				if version.ID == image.CurrentPublicVersion {
-					for _, localImage := range version.LocalImages {
-						images.Images[i].Public = true
-						s.Cache.InsertImage(localImage.ID, localImage.Zone, localImage.Arch, image.Organization.ID, image.Name, image.CurrentPublicVersion)
-					}
-				}
-			}
-		}
-	}
-	values := url.Values{}
-	values.Set("organization", s.Organization)
-	resp, err := s.GetResponsePaginate(s.computeAPI, "images", values)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var OrgaImages ScalewayImages
-
-	if err = json.Unmarshal(body, &OrgaImages); err != nil {
-		return nil, err
-	}
-
-	for _, orgaImage := range OrgaImages.Images {
-		images.Images = append(images.Images, MarketImage{
-			Categories:           []string{"MyImages"},
-			CreationDate:         orgaImage.CreationDate,
-			CurrentPublicVersion: orgaImage.Identifier,
-			ModificationDate:     orgaImage.ModificationDate,
-			Name:                 orgaImage.Name,
-			Public:               false,
-			MarketVersions: MarketVersions{
-				Versions: []MarketVersionDefinition{
-					{
-						CreationDate:     orgaImage.CreationDate,
-						ID:               orgaImage.Identifier,
-						ModificationDate: orgaImage.ModificationDate,
-						MarketLocalImages: MarketLocalImages{
-							LocalImages: []MarketLocalImageDefinition{
-								{
-									Arch: orgaImage.Arch,
-									ID:   orgaImage.Identifier,
-									// TODO: fecth images from ams1 and par1
-									Zone: s.Region,
-								},
-							},
-						},
-					},
-				},
-			},
-		})
-		s.Cache.InsertImage(orgaImage.Identifier, s.Region, orgaImage.Arch, orgaImage.Organization, orgaImage.Name, "")
-	}
-	return &images.Images, nil
-}
-
-// GetImage gets an image from the ScalewayAPI
-func (s *ScalewayAPI) GetImage(imageID string) (*ScalewayImage, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, "images/"+imageID, url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var oneImage ScalewayOneImage
-
-	if err = json.Unmarshal(body, &oneImage); err != nil {
-		return nil, err
-	}
-	// FIXME owner, title
-	s.Cache.InsertImage(oneImage.Image.Identifier, s.Region, oneImage.Image.Arch, oneImage.Image.Organization, oneImage.Image.Name, "")
-	return &oneImage.Image, nil
-}
-
-// DeleteImage deletes a image
-func (s *ScalewayAPI) DeleteImage(imageID string) error {
-	defer s.Cache.RemoveImage(imageID)
-	resp, err := s.DeleteResponse(s.computeAPI, fmt.Sprintf("images/%s", imageID))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if _, err := s.handleHTTPError([]int{http.StatusNoContent}, resp); err != nil {
-		return err
-	}
-	return nil
-}
-
-// DeleteSnapshot deletes a snapshot
-func (s *ScalewayAPI) DeleteSnapshot(snapshotID string) error {
-	defer s.Cache.RemoveSnapshot(snapshotID)
-	resp, err := s.DeleteResponse(s.computeAPI, fmt.Sprintf("snapshots/%s", snapshotID))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if _, err := s.handleHTTPError([]int{http.StatusNoContent}, resp); err != nil {
-		return err
-	}
-	return nil
-}
-
-// DeleteVolume deletes a volume
-func (s *ScalewayAPI) DeleteVolume(volumeID string) error {
-	defer s.Cache.RemoveVolume(volumeID)
-	resp, err := s.DeleteResponse(s.computeAPI, fmt.Sprintf("volumes/%s", volumeID))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if _, err := s.handleHTTPError([]int{http.StatusNoContent}, resp); err != nil {
-		return err
-	}
-	return nil
-}
-
-// GetSnapshots gets the list of snapshots from the ScalewayAPI
-func (s *ScalewayAPI) GetSnapshots() (*[]ScalewaySnapshot, error) {
-	query := url.Values{}
-	s.Cache.ClearSnapshots()
-
-	resp, err := s.GetResponsePaginate(s.computeAPI, "snapshots", query)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var snapshots ScalewaySnapshots
-
-	if err = json.Unmarshal(body, &snapshots); err != nil {
-		return nil, err
-	}
-	for _, snapshot := range snapshots.Snapshots {
-		// FIXME region, arch, owner, title
-		s.Cache.InsertSnapshot(snapshot.Identifier, "", "", snapshot.Organization, snapshot.Name)
-	}
-	return &snapshots.Snapshots, nil
-}
-
-// GetSnapshot gets a snapshot from the ScalewayAPI
-func (s *ScalewayAPI) GetSnapshot(snapshotID string) (*ScalewaySnapshot, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, "snapshots/"+snapshotID, url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var oneSnapshot ScalewayOneSnapshot
-
-	if err = json.Unmarshal(body, &oneSnapshot); err != nil {
-		return nil, err
-	}
-	// FIXME region, arch, owner, title
-	s.Cache.InsertSnapshot(oneSnapshot.Snapshot.Identifier, "", "", oneSnapshot.Snapshot.Organization, oneSnapshot.Snapshot.Name)
-	return &oneSnapshot.Snapshot, nil
-}
-
-// GetVolumes gets the list of volumes from the ScalewayAPI
-func (s *ScalewayAPI) GetVolumes() (*[]ScalewayVolume, error) {
-	query := url.Values{}
-	s.Cache.ClearVolumes()
-
-	resp, err := s.GetResponsePaginate(s.computeAPI, "volumes", query)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-
-	var volumes ScalewayVolumes
-
-	if err = json.Unmarshal(body, &volumes); err != nil {
-		return nil, err
-	}
-	for _, volume := range volumes.Volumes {
-		// FIXME region, arch, owner, title
-		s.Cache.InsertVolume(volume.Identifier, "", "", volume.Organization, volume.Name)
-	}
-	return &volumes.Volumes, nil
-}
-
-// GetVolume gets a volume from the ScalewayAPI
-func (s *ScalewayAPI) GetVolume(volumeID string) (*ScalewayVolume, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, "volumes/"+volumeID, url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var oneVolume ScalewayOneVolume
-
-	if err = json.Unmarshal(body, &oneVolume); err != nil {
-		return nil, err
-	}
-	// FIXME region, arch, owner, title
-	s.Cache.InsertVolume(oneVolume.Volume.Identifier, "", "", oneVolume.Volume.Organization, oneVolume.Volume.Name)
-	return &oneVolume.Volume, nil
-}
-
-// GetBootscripts gets the list of bootscripts from the ScalewayAPI
-func (s *ScalewayAPI) GetBootscripts() (*[]ScalewayBootscript, error) {
-	query := url.Values{}
-
-	s.Cache.ClearBootscripts()
-	resp, err := s.GetResponsePaginate(s.computeAPI, "bootscripts", query)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var bootscripts ScalewayBootscripts
-
-	if err = json.Unmarshal(body, &bootscripts); err != nil {
-		return nil, err
-	}
-	for _, bootscript := range bootscripts.Bootscripts {
-		// FIXME region, arch, owner, title
-		s.Cache.InsertBootscript(bootscript.Identifier, "", bootscript.Arch, bootscript.Organization, bootscript.Title)
-	}
-	return &bootscripts.Bootscripts, nil
-}
-
-// GetBootscript gets a bootscript from the ScalewayAPI
-func (s *ScalewayAPI) GetBootscript(bootscriptID string) (*ScalewayBootscript, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, "bootscripts/"+bootscriptID, url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var oneBootscript ScalewayOneBootscript
-
-	if err = json.Unmarshal(body, &oneBootscript); err != nil {
-		return nil, err
-	}
-	// FIXME region, arch, owner, title
-	s.Cache.InsertBootscript(oneBootscript.Bootscript.Identifier, "", oneBootscript.Bootscript.Arch, oneBootscript.Bootscript.Organization, oneBootscript.Bootscript.Title)
-	return &oneBootscript.Bootscript, nil
-}
-
-// GetUserdatas gets list of userdata for a server
-func (s *ScalewayAPI) GetUserdatas(serverID string, metadata bool) (*ScalewayUserdatas, error) {
-	var uri, endpoint string
-
-	endpoint = s.computeAPI
-	if metadata {
-		uri = "/user_data"
-		endpoint = MetadataAPI
-	} else {
-		uri = fmt.Sprintf("servers/%s/user_data", serverID)
-	}
-
-	resp, err := s.GetResponsePaginate(endpoint, uri, url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var userdatas ScalewayUserdatas
-
-	if err = json.Unmarshal(body, &userdatas); err != nil {
-		return nil, err
-	}
-	return &userdatas, nil
-}
-
-func (s *ScalewayUserdata) String() string {
-	return string(*s)
-}
-
-// GetUserdata gets a specific userdata for a server
-func (s *ScalewayAPI) GetUserdata(serverID, key string, metadata bool) (*ScalewayUserdata, error) {
-	var uri, endpoint string
-
-	endpoint = s.computeAPI
-	if metadata {
-		uri = fmt.Sprintf("/user_data/%s", key)
-		endpoint = MetadataAPI
-	} else {
-		uri = fmt.Sprintf("servers/%s/user_data/%s", serverID, key)
-	}
-
-	var err error
-	resp, err := s.GetResponsePaginate(endpoint, uri, url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("no such user_data %q (%d)", key, resp.StatusCode)
-	}
-	var data ScalewayUserdata
-	data, err = ioutil.ReadAll(resp.Body)
-	return &data, err
-}
-
-// PatchUserdata sets a user data
-func (s *ScalewayAPI) PatchUserdata(serverID, key string, value []byte, metadata bool) error {
-	var resource, endpoint string
-
-	endpoint = s.computeAPI
-	if metadata {
-		resource = fmt.Sprintf("/user_data/%s", key)
-		endpoint = MetadataAPI
-	} else {
-		resource = fmt.Sprintf("servers/%s/user_data/%s", serverID, key)
-	}
-
-	uri := fmt.Sprintf("%s/%s", strings.TrimRight(endpoint, "/"), resource)
-	payload := new(bytes.Buffer)
-	payload.Write(value)
-
-	req, err := http.NewRequest("PATCH", uri, payload)
-	if err != nil {
-		return err
-	}
-
-	req.Header.Set("X-Auth-Token", s.Token)
-	req.Header.Set("Content-Type", "text/plain")
-	req.Header.Set("User-Agent", s.userAgent)
-
-	s.LogHTTP(req)
-
-	resp, err := s.client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode == http.StatusNoContent {
-		return nil
-	}
-
-	return fmt.Errorf("cannot set user_data (%d)", resp.StatusCode)
-}
-
-// DeleteUserdata deletes a server user_data
-func (s *ScalewayAPI) DeleteUserdata(serverID, key string, metadata bool) error {
-	var url, endpoint string
-
-	endpoint = s.computeAPI
-	if metadata {
-		url = fmt.Sprintf("/user_data/%s", key)
-		endpoint = MetadataAPI
-	} else {
-		url = fmt.Sprintf("servers/%s/user_data/%s", serverID, key)
-	}
-
-	resp, err := s.DeleteResponse(endpoint, url)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = s.handleHTTPError([]int{http.StatusNoContent}, resp)
-	return err
 }
 
 // GetTasks get the list of tasks from the ScalewayAPI
@@ -2030,24 +1275,6 @@ func (s *ScalewayAPI) GetDashboard() (*ScalewayDashboard, error) {
 	return &dashboard.Dashboard, nil
 }
 
-// GetServerID returns exactly one server matching
-func (s *ScalewayAPI) GetServerID(needle string) (string, error) {
-	// Parses optional type prefix, i.e: "server:name" -> "name"
-	_, needle = parseNeedle(needle)
-
-	servers, err := s.ResolveServer(needle)
-	if err != nil {
-		return "", fmt.Errorf("Unable to resolve server %s: %s", needle, err)
-	}
-	if len(servers) == 1 {
-		return servers[0].Identifier, nil
-	}
-	if len(servers) == 0 {
-		return "", fmt.Errorf("No such server: %s", needle)
-	}
-	return "", showResolverResults(needle, servers)
-}
-
 func showResolverResults(needle string, results ScalewayResolverResults) error {
 	w := tabwriter.NewWriter(os.Stderr, 20, 1, 3, ' ', 0)
 	defer w.Flush()
@@ -2060,246 +1287,6 @@ func showResolverResults(needle string, results ScalewayResolverResults) error {
 		fmt.Fprintf(w, "- %s\t%s\t%s\t%s\t%s\n", result.TruncIdentifier(), result.CodeName(), result.Name, result.Region, result.Arch)
 	}
 	return fmt.Errorf("Too many candidates for %s (%d)", needle, len(results))
-}
-
-// GetVolumeID returns exactly one volume matching
-func (s *ScalewayAPI) GetVolumeID(needle string) (string, error) {
-	// Parses optional type prefix, i.e: "volume:name" -> "name"
-	_, needle = parseNeedle(needle)
-
-	volumes, err := s.ResolveVolume(needle)
-	if err != nil {
-		return "", fmt.Errorf("Unable to resolve volume %s: %s", needle, err)
-	}
-	if len(volumes) == 1 {
-		return volumes[0].Identifier, nil
-	}
-	if len(volumes) == 0 {
-		return "", fmt.Errorf("No such volume: %s", needle)
-	}
-	return "", showResolverResults(needle, volumes)
-}
-
-// GetSnapshotID returns exactly one snapshot matching
-func (s *ScalewayAPI) GetSnapshotID(needle string) (string, error) {
-	// Parses optional type prefix, i.e: "snapshot:name" -> "name"
-	_, needle = parseNeedle(needle)
-
-	snapshots, err := s.ResolveSnapshot(needle)
-	if err != nil {
-		return "", fmt.Errorf("Unable to resolve snapshot %s: %s", needle, err)
-	}
-	if len(snapshots) == 1 {
-		return snapshots[0].Identifier, nil
-	}
-	if len(snapshots) == 0 {
-		return "", fmt.Errorf("No such snapshot: %s", needle)
-	}
-	return "", showResolverResults(needle, snapshots)
-}
-
-// FilterImagesByArch removes entry that doesn't match with architecture
-func FilterImagesByArch(res ScalewayResolverResults, arch string) (ret ScalewayResolverResults) {
-	if arch == "*" {
-		return res
-	}
-	for _, result := range res {
-		if result.Arch == arch {
-			ret = append(ret, result)
-		}
-	}
-	return
-}
-
-// FilterImagesByRegion removes entry that doesn't match with region
-func FilterImagesByRegion(res ScalewayResolverResults, region string) (ret ScalewayResolverResults) {
-	if region == "*" {
-		return res
-	}
-	for _, result := range res {
-		if result.Region == region {
-			ret = append(ret, result)
-		}
-	}
-	return
-}
-
-// GetImageID returns exactly one image matching
-func (s *ScalewayAPI) GetImageID(needle, arch string) (*ScalewayImageIdentifier, error) {
-	// Parses optional type prefix, i.e: "image:name" -> "name"
-	_, needle = parseNeedle(needle)
-
-	images, err := s.ResolveImage(needle)
-	if err != nil {
-		return nil, fmt.Errorf("Unable to resolve image %s: %s", needle, err)
-	}
-	images = FilterImagesByArch(images, arch)
-	images = FilterImagesByRegion(images, s.Region)
-	if len(images) == 1 {
-		return &ScalewayImageIdentifier{
-			Identifier: images[0].Identifier,
-			Arch:       images[0].Arch,
-			// FIXME region, owner hardcoded
-			Region: images[0].Region,
-			Owner:  "",
-		}, nil
-	}
-	if len(images) == 0 {
-		return nil, fmt.Errorf("No such image (zone %s, arch %s) : %s", s.Region, arch, needle)
-	}
-	return nil, showResolverResults(needle, images)
-}
-
-// GetSecurityGroups returns a ScalewaySecurityGroups
-func (s *ScalewayAPI) GetSecurityGroups() (*ScalewayGetSecurityGroups, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, "security_groups", url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var securityGroups ScalewayGetSecurityGroups
-
-	if err = json.Unmarshal(body, &securityGroups); err != nil {
-		return nil, err
-	}
-	return &securityGroups, nil
-}
-
-// GetSecurityGroupRules returns a ScalewaySecurityGroupRules
-func (s *ScalewayAPI) GetSecurityGroupRules(groupID string) (*ScalewayGetSecurityGroupRules, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, fmt.Sprintf("security_groups/%s/rules", groupID), url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var securityGroupRules ScalewayGetSecurityGroupRules
-
-	if err = json.Unmarshal(body, &securityGroupRules); err != nil {
-		return nil, err
-	}
-	return &securityGroupRules, nil
-}
-
-// GetASecurityGroupRule returns a ScalewaySecurityGroupRule
-func (s *ScalewayAPI) GetASecurityGroupRule(groupID string, rulesID string) (*ScalewayGetSecurityGroupRule, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, fmt.Sprintf("security_groups/%s/rules/%s", groupID, rulesID), url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var securityGroupRules ScalewayGetSecurityGroupRule
-
-	if err = json.Unmarshal(body, &securityGroupRules); err != nil {
-		return nil, err
-	}
-	return &securityGroupRules, nil
-}
-
-// GetASecurityGroup returns a ScalewaySecurityGroup
-func (s *ScalewayAPI) GetASecurityGroup(groupsID string) (*ScalewayGetSecurityGroup, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, fmt.Sprintf("security_groups/%s", groupsID), url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var securityGroups ScalewayGetSecurityGroup
-
-	if err = json.Unmarshal(body, &securityGroups); err != nil {
-		return nil, err
-	}
-	return &securityGroups, nil
-}
-
-// PostSecurityGroup posts a group on a server
-func (s *ScalewayAPI) PostSecurityGroup(group ScalewayNewSecurityGroup) error {
-	resp, err := s.PostResponse(s.computeAPI, "security_groups", group)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = s.handleHTTPError([]int{http.StatusCreated}, resp)
-	return err
-}
-
-// PostSecurityGroupRule posts a rule on a server
-func (s *ScalewayAPI) PostSecurityGroupRule(SecurityGroupID string, rules ScalewayNewSecurityGroupRule) error {
-	resp, err := s.PostResponse(s.computeAPI, fmt.Sprintf("security_groups/%s/rules", SecurityGroupID), rules)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = s.handleHTTPError([]int{http.StatusCreated}, resp)
-	return err
-}
-
-// DeleteSecurityGroup deletes a SecurityGroup
-func (s *ScalewayAPI) DeleteSecurityGroup(securityGroupID string) error {
-	resp, err := s.DeleteResponse(s.computeAPI, fmt.Sprintf("security_groups/%s", securityGroupID))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = s.handleHTTPError([]int{http.StatusNoContent}, resp)
-	return err
-}
-
-// PutSecurityGroup updates a SecurityGroup
-func (s *ScalewayAPI) PutSecurityGroup(group ScalewayUpdateSecurityGroup, securityGroupID string) error {
-	resp, err := s.PutResponse(s.computeAPI, fmt.Sprintf("security_groups/%s", securityGroupID), group)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = s.handleHTTPError([]int{http.StatusOK}, resp)
-	return err
-}
-
-// PutSecurityGroupRule updates a SecurityGroupRule
-func (s *ScalewayAPI) PutSecurityGroupRule(rules ScalewayNewSecurityGroupRule, securityGroupID, RuleID string) error {
-	resp, err := s.PutResponse(s.computeAPI, fmt.Sprintf("security_groups/%s/rules/%s", securityGroupID, RuleID), rules)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = s.handleHTTPError([]int{http.StatusOK}, resp)
-	return err
-}
-
-// DeleteSecurityGroupRule deletes a SecurityGroupRule
-func (s *ScalewayAPI) DeleteSecurityGroupRule(SecurityGroupID, RuleID string) error {
-	resp, err := s.DeleteResponse(s.computeAPI, fmt.Sprintf("security_groups/%s/rules/%s", SecurityGroupID, RuleID))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	_, err = s.handleHTTPError([]int{http.StatusNoContent}, resp)
-	return err
 }
 
 // GetContainers returns a ScalewayGetContainers
@@ -2342,123 +1329,6 @@ func (s *ScalewayAPI) GetContainerDatas(container string) (*ScalewayGetContainer
 	return &datas, nil
 }
 
-// GetIPS returns a ScalewayGetIPS
-func (s *ScalewayAPI) GetIPS() (*ScalewayGetIPS, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, "ips", url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var ips ScalewayGetIPS
-
-	if err = json.Unmarshal(body, &ips); err != nil {
-		return nil, err
-	}
-	return &ips, nil
-}
-
-// NewIP returns a new IP
-func (s *ScalewayAPI) NewIP() (*ScalewayGetIP, error) {
-	var orga struct {
-		Organization string `json:"organization"`
-	}
-	orga.Organization = s.Organization
-	resp, err := s.PostResponse(s.computeAPI, "ips", orga)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusCreated}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var ip ScalewayGetIP
-
-	if err = json.Unmarshal(body, &ip); err != nil {
-		return nil, err
-	}
-	return &ip, nil
-}
-
-// AttachIP attachs an IP to a server
-func (s *ScalewayAPI) AttachIP(ipID, serverID string) error {
-	var update struct {
-		Address      string  `json:"address"`
-		ID           string  `json:"id"`
-		Reverse      *string `json:"reverse"`
-		Organization string  `json:"organization"`
-		Server       string  `json:"server"`
-	}
-
-	ip, err := s.GetIP(ipID)
-	if err != nil {
-		return err
-	}
-	update.Address = ip.IP.Address
-	update.ID = ip.IP.ID
-	update.Organization = ip.IP.Organization
-	update.Server = serverID
-	resp, err := s.PutResponse(s.computeAPI, fmt.Sprintf("ips/%s", ipID), update)
-	if err != nil {
-		return err
-	}
-	_, err = s.handleHTTPError([]int{http.StatusOK}, resp)
-	return err
-}
-
-// DetachIP detaches an IP from a server
-func (s *ScalewayAPI) DetachIP(ipID string) error {
-	ip, err := s.GetIP(ipID)
-	if err != nil {
-		return err
-	}
-	ip.IP.Server = nil
-	resp, err := s.PutResponse(s.computeAPI, fmt.Sprintf("ips/%s", ipID), ip.IP)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusOK}, resp)
-	return err
-}
-
-// DeleteIP deletes an IP
-func (s *ScalewayAPI) DeleteIP(ipID string) error {
-	resp, err := s.DeleteResponse(s.computeAPI, fmt.Sprintf("ips/%s", ipID))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusNoContent}, resp)
-	return err
-}
-
-// GetIP returns a ScalewayGetIP
-func (s *ScalewayAPI) GetIP(ipID string) (*ScalewayGetIP, error) {
-	resp, err := s.GetResponsePaginate(s.computeAPI, fmt.Sprintf("ips/%s", ipID), url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var ip ScalewayGetIP
-
-	if err = json.Unmarshal(body, &ip); err != nil {
-		return nil, err
-	}
-	return &ip, nil
-}
-
 // GetQuotas returns a ScalewayGetQuotas
 func (s *ScalewayAPI) GetQuotas() (*ScalewayGetQuotas, error) {
 	resp, err := s.GetResponsePaginate(AccountAPI, fmt.Sprintf("organizations/%s/quotas", s.Organization), url.Values{})
@@ -2477,25 +1347,6 @@ func (s *ScalewayAPI) GetQuotas() (*ScalewayGetQuotas, error) {
 		return nil, err
 	}
 	return &quotas, nil
-}
-
-// GetBootscriptID returns exactly one bootscript matching
-func (s *ScalewayAPI) GetBootscriptID(needle, arch string) (string, error) {
-	// Parses optional type prefix, i.e: "bootscript:name" -> "name"
-	_, needle = parseNeedle(needle)
-
-	bootscripts, err := s.ResolveBootscript(needle)
-	if err != nil {
-		return "", fmt.Errorf("Unable to resolve bootscript %s: %s", needle, err)
-	}
-	bootscripts.FilterByArch(arch)
-	if len(bootscripts) == 1 {
-		return bootscripts[0].Identifier, nil
-	}
-	if len(bootscripts) == 0 {
-		return "", fmt.Errorf("No such bootscript: %s", needle)
-	}
-	return "", showResolverResults(needle, bootscripts)
 }
 
 func rootNetDial(network, addr string) (net.Conn, error) {
@@ -2532,214 +1383,6 @@ func rootNetDial(network, addr string) (net.Conn, error) {
 // SetPassword register the password
 func (s *ScalewayAPI) SetPassword(password string) {
 	s.password = password
-}
-
-// GetMarketPlaceImages returns images from marketplace
-func (s *ScalewayAPI) GetMarketPlaceImages(uuidImage string) (*MarketImages, error) {
-	resp, err := s.GetResponsePaginate(MarketplaceAPI, fmt.Sprintf("images/%s", uuidImage), url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var ret MarketImages
-
-	if uuidImage != "" {
-		ret.Images = make([]MarketImage, 1)
-
-		var img MarketImage
-
-		if err = json.Unmarshal(body, &img); err != nil {
-			return nil, err
-		}
-		ret.Images[0] = img
-	} else {
-		if err = json.Unmarshal(body, &ret); err != nil {
-			return nil, err
-		}
-	}
-	return &ret, nil
-}
-
-// GetMarketPlaceImageVersions returns image version
-func (s *ScalewayAPI) GetMarketPlaceImageVersions(uuidImage, uuidVersion string) (*MarketVersions, error) {
-	resp, err := s.GetResponsePaginate(MarketplaceAPI, fmt.Sprintf("images/%v/versions/%s", uuidImage, uuidVersion), url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var ret MarketVersions
-
-	if uuidImage != "" {
-		var version MarketVersion
-		ret.Versions = make([]MarketVersionDefinition, 1)
-
-		if err = json.Unmarshal(body, &version); err != nil {
-			return nil, err
-		}
-		ret.Versions[0] = version.Version
-	} else {
-		if err = json.Unmarshal(body, &ret); err != nil {
-			return nil, err
-		}
-	}
-	return &ret, nil
-}
-
-// GetMarketPlaceImageCurrentVersion return the image current version
-func (s *ScalewayAPI) GetMarketPlaceImageCurrentVersion(uuidImage string) (*MarketVersion, error) {
-	resp, err := s.GetResponsePaginate(MarketplaceAPI, fmt.Sprintf("images/%v/versions/current", uuidImage), url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var ret MarketVersion
-
-	if err = json.Unmarshal(body, &ret); err != nil {
-		return nil, err
-	}
-	return &ret, nil
-}
-
-// GetMarketPlaceLocalImages returns images from local region
-func (s *ScalewayAPI) GetMarketPlaceLocalImages(uuidImage, uuidVersion, uuidLocalImage string) (*MarketLocalImages, error) {
-	resp, err := s.GetResponsePaginate(MarketplaceAPI, fmt.Sprintf("images/%v/versions/%s/local_images/%s", uuidImage, uuidVersion, uuidLocalImage), url.Values{})
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	body, err := s.handleHTTPError([]int{http.StatusOK}, resp)
-	if err != nil {
-		return nil, err
-	}
-	var ret MarketLocalImages
-	if uuidLocalImage != "" {
-		var localImage MarketLocalImage
-		ret.LocalImages = make([]MarketLocalImageDefinition, 1)
-
-		if err = json.Unmarshal(body, &localImage); err != nil {
-			return nil, err
-		}
-		ret.LocalImages[0] = localImage.LocalImages
-	} else {
-		if err = json.Unmarshal(body, &ret); err != nil {
-			return nil, err
-		}
-	}
-	return &ret, nil
-}
-
-// PostMarketPlaceImage adds new image
-func (s *ScalewayAPI) PostMarketPlaceImage(images MarketImage) error {
-	resp, err := s.PostResponse(MarketplaceAPI, "images/", images)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusAccepted}, resp)
-	return err
-}
-
-// PostMarketPlaceImageVersion adds new image version
-func (s *ScalewayAPI) PostMarketPlaceImageVersion(uuidImage string, version MarketVersion) error {
-	resp, err := s.PostResponse(MarketplaceAPI, fmt.Sprintf("images/%v/versions", uuidImage), version)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusAccepted}, resp)
-	return err
-}
-
-// PostMarketPlaceLocalImage adds new local image
-func (s *ScalewayAPI) PostMarketPlaceLocalImage(uuidImage, uuidVersion, uuidLocalImage string, local MarketLocalImage) error {
-	resp, err := s.PostResponse(MarketplaceAPI, fmt.Sprintf("images/%v/versions/%s/local_images/%v", uuidImage, uuidVersion, uuidLocalImage), local)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusAccepted}, resp)
-	return err
-}
-
-// PutMarketPlaceImage updates image
-func (s *ScalewayAPI) PutMarketPlaceImage(uudiImage string, images MarketImage) error {
-	resp, err := s.PutResponse(MarketplaceAPI, fmt.Sprintf("images/%v", uudiImage), images)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusOK}, resp)
-	return err
-}
-
-// PutMarketPlaceImageVersion updates image version
-func (s *ScalewayAPI) PutMarketPlaceImageVersion(uuidImage, uuidVersion string, version MarketVersion) error {
-	resp, err := s.PutResponse(MarketplaceAPI, fmt.Sprintf("images/%v/versions/%v", uuidImage, uuidVersion), version)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusOK}, resp)
-	return err
-}
-
-// PutMarketPlaceLocalImage updates local image
-func (s *ScalewayAPI) PutMarketPlaceLocalImage(uuidImage, uuidVersion, uuidLocalImage string, local MarketLocalImage) error {
-	resp, err := s.PostResponse(MarketplaceAPI, fmt.Sprintf("images/%v/versions/%s/local_images/%v", uuidImage, uuidVersion, uuidLocalImage), local)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusOK}, resp)
-	return err
-}
-
-// DeleteMarketPlaceImage deletes image
-func (s *ScalewayAPI) DeleteMarketPlaceImage(uudImage string) error {
-	resp, err := s.DeleteResponse(MarketplaceAPI, fmt.Sprintf("images/%v", uudImage))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusNoContent}, resp)
-	return err
-}
-
-// DeleteMarketPlaceImageVersion delete image version
-func (s *ScalewayAPI) DeleteMarketPlaceImageVersion(uuidImage, uuidVersion string) error {
-	resp, err := s.DeleteResponse(MarketplaceAPI, fmt.Sprintf("images/%v/versions/%v", uuidImage, uuidVersion))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusNoContent}, resp)
-	return err
-}
-
-// DeleteMarketPlaceLocalImage deletes local image
-func (s *ScalewayAPI) DeleteMarketPlaceLocalImage(uuidImage, uuidVersion, uuidLocalImage string) error {
-	resp, err := s.DeleteResponse(MarketplaceAPI, fmt.Sprintf("images/%v/versions/%s/local_images/%v", uuidImage, uuidVersion, uuidLocalImage))
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	_, err = s.handleHTTPError([]int{http.StatusNoContent}, resp)
-	return err
 }
 
 // ResolveTTYUrl return an URL to get a tty
